requirements.txt
asgiref==3.3.1
Django==3.1.7
django-multiselectfield==0.1.12
Pillow==8.2.0
python-slugify==4.0.1
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
Unidecode==1.2.0
